Add something

+++ Lab 3/sample_input3.c
@@ -4,6 +4,6 @@ int main()
 	b=3;
 	float c=4.5;
 	c=6.5;
-	double d=5.44;
-	double e=d+9.0-4.0/2.0;
+	float d=5.44;
+	float e=d+9.0-4.0/2.0;
 }
